Split RAS_ITexVert and RAS_IDisplayArray in separated files.
